UPSTREAM: arm64: dts: mediatek: mt8173-elm-hana: Add vdd-supply to second source trackpad
The Hana device has a second source option trackpad, but it is missing
its regulator supply. It only works because the regulator is marked as
always-on.

Add the regulator supply, but leave out the post-power-on delay. Instead,
document the post-power-on delay along with the reason for not adding
it in a comment.

Fixes: 689b937bedde ("arm64: dts: mediatek: add mt8173 elm and hana board")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20241018082001.1296963-1-wenst@chromium.org
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit f766fae08f6a2eaeb45d8d2c053724c91526835c)

BUG=b:149607803
TEST=boot elm on 5.4 with UI

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Ie4456ed2bc0a33a3b07cb62ddc8985c723ed6bae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6058697
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Divya Chauhan <chauhandivya@google.com>
Reviewed-by: Divya Chauhan <chauhandivya@google.com>