CHROMIUM: iwl7000: mld: improve DEFINE_SIMPLE_CANCELLATION
This macro generates a simple cancellation function for a async
notification.
Currently, it receives the size of the id member within the notification
struct.
But the current implementation doesn't work for u8.
Change the macro to use _Generic in order to determine the type,
which allows us to remve the size parameter and makes it work for u8
too.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ib3ad490b43b326500891b1caeab9e2693aaaeb26
iwl7000-tree: 9eea7b2157870db9c197a43147b8be43375c42c0