CHROMIUM: dma-buf: heaps: restricted_heap: Return EACCES instead of EPERM
EACCES A file descriptor refers to a non-regular file.  Or a file
mapping was requested, but fd is not open for reading.  Or MAP_SHARED
was requested and PROT_WRITE is set, but fd is not open in read/write
(O_RDWR) mode.  Or PROT_WRITE is set, but the file is append-only

BUG=b:384013833
TEST=emerge-rauru chromeos-kernel-6_6; boot to shell
UPSTREAM-TASK=b:387623307

Change-Id: I5282c6f579921a3d88a4dbe8d8a7aaaad02f9e1e
Signed-off-by: Yong Wu <yong.wu@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6073985
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>