CHROMIUM: iwl7000: mld: don't cycle deflink when FW is dead
When the drv_unassign_chanctx is called for deflink (non-MLO),
and we are disassociating, we cycle the deflink (remove and re-add it).
This is done in order to cleanup deflink from the old connection data.

But if we failed in one of the stages of the disassoc flow because of a
fw error, there is no point in doing this
(because FW is dead so all the commands will fail).

And if we do that, the command to add the link to the fw fails, so we
unlink link[0] from deflink, and then it will not necessarily be linked
again (for example if we are doing reprobe and not sw reset).

OTOH, link[0] is assumed to be valid for the lifetime of the vif
(in non-MLO) and then later in the reprobe, when the vif is being
removed, we will have a NULL pointer dereference.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ib88904eea299ff538378f7ec0b0285eb933438be
iwl7000-tree: 42fc3b2b0b82535b693f1cc84a36d9cfc0f0b6f7