FROMGIT: media: uvcvideo: Only save async fh if success
Now we keep a reference to the active fh for any call to uvc_ctrl_set,
regardless if it is an actual set or if it is a just a try or if the
device refused the operation.

We should only keep the file handle if the device actually accepted
applying the operation.

Cc: stable@vger.kernel.org
Fixes: e5225c820c05 ("media: uvcvideo: Send a control event when a Control Change interrupt arrives")
Suggested-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Link: https://lore.kernel.org/r/20241203-uvc-fix-async-v6-1-26c867231118@chromium.org
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit d9fecd096f67a4469536e040a8a10bbfb665918b
 https://gitlab.freedesktop.org/linux-media/media-committers.git next)

BUG=b:381229714
TEST=Compile test

Change-Id: I8c1e686a0e014c0320388bcf2c8ab30e7e2ee62d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6111688
Commit-Queue: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>
Tested-by: Ricardo Ribalda <ribalda@chromium.org>
Auto-Submit: Ricardo Ribalda <ribalda@chromium.org>