CHROMIUM: iwl7000: mac80211: don't flush non-uploaded STAs
If STA state is pre-moved to AUTHORIZED (such as in IBSS
scenarios) and insertion fails, the station is freed.
In this case, the driver never knew about the station,
so trying to flush it is unexpected and may crash.

Check if the sta was uploaded to the driver before and
fix this.

Change-Id: I491bbcccc46f835ade07df0640a75f6ed92f20a3
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4a5e078093d0d6ba81068ba7a2b0c2ebd5858cd4