CHROMIUM: fuse: add hung_request watchdog
Add a hung_task watchdog to detect stalled request so that
we can terminate the connection and avoid locking up the
user-space.

The core part is taken from [1], stripped down to bare
minimum that we need, and converted to a task-based
watchdog.

[1] https://lore.kernel.org/all/20241114191332.669127-3-joannelkoong@gmail.com/

UPSTREAM-TASK=b:381968048
BUG=b:373731719
TEST=accessed gdrive via Files app on dedede
     kill -STOP $(pidof drivefs)  .. long wait ..
     kill -CONT $(pidof drivefs)
     both before and during fuse req processing
     under lockdep, kasan

Change-Id: I3b0c39af61af89a5e91ce2e21b39899dfb7364a4
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6076182
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>