CHROMIUM: Increase DPM watchdog timeout to 15 seconds
In <https://crrev.com/c/5341615> we changfed the DPM watchdog from 60
seconds to 10. This is causing problems. Change it to 15 seconds.

Some background here:
* On most recent systems the EC has a sleep hang detector. When the AP
  starts going to suspend it tells the EC about it's plan. The EC then
  monitors a GPIO that is supposed to assert when the AP successfully
  gets into a low power mode. If the EC doesn't see this GPIO assert
  after 10 seconds then it's a "sleep hang". When the EC sees this
  sleep hang it asserts a wakeup signal to the AP. The userspace power
  manager in ChromeOS knows about this case and will try to suspend
  again. If userspace sees too many failed suspends in a row it will
  try to shut the system down.
* On newer ECs the sleep hang detector has apparently been improved to
  have a "hard sleep hang dector". After the EC wakes the AP for a
  soft sleep hang then it expects the AP to tell it that it's
  awake. If the EC never sees it then it's a hard sleep hang and the
  EC will try to reboot. See https://crrev.com/c/4749162
* On even newer ECs, there's even more complex logic. See
  https://crrev.com/c/5369987

In general, panicking the AP is pretty disruptive and should be a last
resort. Better than doing a panic() is to try first to abort the
suspend / wake the system back up so that it can try to suspend
again. There's a reasonable chance that retrying the suspend will
work. If it doesn't there is already logic in userspace to eventually
give up and shut the system down.

Eventually, it seems like it would be a good idea to make the DPM
watchdog try to abort the suspend / wake the system up instead of
jumping straight to a panic(). While considering that, let's at least
set the timeout to 15 seconds. It has been found that there are
several timeouts in the kernel tuned at around 10 seconds. In at least
one case after the timeout expired then the system would have gone to
suspend. In another case it's believed that the system would have
woken up and then a future suspend might have worked. Because of the
short DPM timeout both were now causing kernel panic(). While both
cases were bugs that should be addressed, doing a panic was worse for
the end user.

BUG=b:382269699
TEST=suspend/resume

Change-Id: If1f778730dd167830c0698d2d2fdbefc61a7c633
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6168887
Reviewed-by: Boris Mittelberg <bmbm@google.com>