CHROMIUM: iwl7000: mld: cancel an unsuccessful ROC that started
If the ROC started unsuccessfully, we will report expired to mac80211
which will retry immediately. This will make the firmware crash because
the ROC is running even if it is unsuccessful.
Remove the existing ROC first and only then report expired to mac80211
to fix this issue.

Change-Id: I07b700849044d6a24dd203fc713e71783c21faa1
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 3b9cf62b468e2e55ca7aa6d5ee3ec3758492bb75