CHROMIUM: iwl7000: mld: Store firmware RFI state in MLD global data
Stored the firmware RFI state in the MLD global data structure.
This allows us to determine whether the RFI state is good or not
in the firmware. With this information, we can now send and
receive the DDR table to/from the firmware at the appropriate times.

Signed-off-by: Anjaneyulu <pagadala.yesu.anjaneyulu@intel.com>
Change-Id: I823b794d8bedd79ef471d27c817e2255af5f7bd7
iwl7000-tree: 77c8f188be760807600194b8fb75d7ce6296b2ab