CHROMIUM: iwl7000: mld: Check RFI support before retrieving frequency table
Added a check to determine if at least one of the RFI features are
supported before proceeding with the retrieval of the RFI frequency
table. This ensures that the operation is only attempted if the
DDR/DLVR/DESENSE feature is supported, preventing unnecessary errors.

Signed-off-by: Anjaneyulu <pagadala.yesu.anjaneyulu@intel.com>
Change-Id: If3ab06e7edb756837021b1f8ec0a694ac895f92b
iwl7000-tree: ef6ebd4f14497f727554b7b0e596933a2f085f19