CHROMIUM: iwl7000: mld: purge the async handler list on stop
When drv_stop is called, we do the following:
1. flush the async_handker_wk
2. stop the HW
3. cancel the async_handler_wk, in case it was re-scheduled between 1
   and 2.

But, if it was re-scheduled between 1 and 2, and we canceled it in 3,
the async_handler_list will not be purged, and its mempry won't be
freed, resulting in a memory leak.

Fix this by purging the list after the cancellation of the worker.
While at it, improve the comments.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I70fc6d1531d26b2b32eab188b4f57c1663926ea9
iwl7000-tree: d5392e3663772dcb768e7233a12a231df88379be