CHROMIUM: iwl7000: mld: add minimum throughput criteria for EMLSR
Only enter EMLSR if a certain minimum throughput has been reached and is
kept for each notification period (which is 5 seconds). Also verify that
both links are actually used and exit EMLSR again (but not block it) if
the secondary link is not used.

Change-Id: Ifbd58be36fe9c747065ce27de29b42ee4445949c
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
iwl7000-tree: 31ad29ddc16b1ab6a067f99ef58ec44eca276f15