CHROMIUM: iwl7000: mld: don't handle the Noa data for a SoftAP vif
When we get a probe resp data notif from the firmware, we do two things:

* update the beacon count down in case of CSA
* update the NoA P2P attribute

The first one needs a mac80211 that is not MLD ready, the latter is
relevant for P2P GO we can't be MLD either.

Exit early if we're not a P2P GO, and ensure we don't have an MLD vif in
hand.

Change-Id: I870d3d993c5799dc9e7b2287079dd3500b35d099
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 3e31fb54550aafb2e63aece219d569d906c4b172