CHROMIUM: iwl7000: mac80211: aggregation: remove deflink accesses for MLO
If a station has connected with MLO (as indicated by valid_links
being non-zero, even if that may have just a single bit set), it
necessarily supports EHT/aggregation, so we don't need to check
the deflink for those cases.

Add conditions so we can support removing the link it/we used to
associate on.

Note that we still use the statistics in the deflink, but that's
a whole different story we will need to address separately.

Change-Id: Id2c2dfbbe7451cc900ed88c5a81b33c55b4ab1cf
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 55061340d03e2e6379c4009d4707bd1c3ab6f637