CHROMIUM: iwl7000: mld: add iwl_mld_clear_stats_in_fw
When we connect, we need to tell the FW to clear its statistics.
(and maybe also clear ours, but that is still a TODO).

Currently we do it by calling iwl_mld_request_fw_stats.
But this one is also waiting for the notifications to arrive
and flushes the async handlers wk in order to get the statistics
handled in place (which we don't really need in the reset case).

Separate the clear case to another function which only tells the FW to
clear its statistics.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I32fe5a0ff6728a1c4aae58e15cca7886a90efa7f
iwl7000-tree: b4385ea4134bd75cc4f9aeadeadc027bea309ce4