CHROMIUM: iwl7000: mld: add iwl_mld_clear_stats_in_fw
When we connect, we need to tell the FW to clear its statistics.
(and maybe also clear ours, but that is still a TODO).

Currently we do it by calling iwl_mld_request_fw_stats.
But this one is also waiting for the notifications to arrive
and flushes the async handlers wk in order to get the statistics
handled in place (which we don't really need in the reset case).

Separate the clear case to another function which only tells the FW to
clear its statistics.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Iab8745c92b6daa1027f45d87eadbfd6c132d88f4
iwl7000-tree: f8324daa4cb5ecdc6b6a161da0e7a3b43794e20e