CHROMIUM: iwl7000: mld: avoid flushing on drv_sta_statistics
The different statistics notifications are handled async, but if we want
to make sure we processed it (in order to give mac80211 up-to-date info),
we currently do wiphy_work_flush(async_handler_wk) so the notifications
will get processed.
But since flushing wiphy works turns out to be dangerous (introduces
function re-entrance) we try to avoid it as much as possible.

Instead of flushing, register a handler to the notifications waiter that
simply populates the relevant data from the notifications directly to the
output struct (station_info).

Later when the async handler will run, it'll fully process the
notification.

This makes iwl_mld_link_sta:ave_energy unneeded. Remove it.

Note: we might want to cancel the async handling, for example to avoid
clearing the mpdu counters just because mac80211 requested statistics,
and not because the window ended (which happens when the *periodic*
statistics come).

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I67067b7ac46ac35e6fa6511b0d7eeb7c18f0822a
iwl7000-tree: 27c5adba2e59b32140492415bffb07864d7249ea