CHROMIUM: iwl7000: mld: avoid flushing on drv_sta_statistics
The different statistics notifications are handled async, but if we want
to make sure we processed it (in order to give mac80211 up-to-date info),
we currently do wiphy_work_flush(async_handler_wk) so the notifications
will get processed.
But since flushing wiphy works turns out to be dangerous (introduces
function re-entrance) we try to avoid it as much as possible.

Instead of flushing, register a handler to the notifications waiter that
simply populates the relevant data from the notifications directly to the
output struct (station_info).

Later when the async handler will run, it'll fully process the
notification.

This makes iwl_mld_link_sta:ave_energy unneeded. Remove it.

Note: we might want to cancel the async handling, for example to avoid
clearing the mpdu counters just because mac80211 requested statistics,
and not because the window ended (which happens when the *periodic*
statistics come).

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I93f88e1f8eecc4ace1f163c2dcb8ceafd8cfedfa
iwl7000-tree: 89ff6f84f857a51092e5ec6c1993dc9917118b87