CHROMIUM: iwl7000: mld: don't flush the worker from mac_flush()
We don't really need to flush the worker when we have a flush() request
from mac80211. All we need is to allocate whatever queue was needed.
Note that this means that we'll allocate the queue (stealing the task
from the worker that was scheduled) out of order with respect to other
workers that were added to the wiphy workqueue, but that's fine. We are
flushing anyways...

Allocate the queue in-place if needed (from mac_flush()) and remove the
wiphy_work_flush().

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Change-Id: I7786f80766b13e66ca6ca6b82ef6408425392346
iwl7000-tree: a35128c80f558bb1d4e2097bccd4b24891560741