CHROMIUM: iwl7000: mld: don't flush the worker from mac_flush()
We don't really need to flush the worker when we have a flush() request
from mac80211. All we need is to allocate whatever queue was needed.
Note that this means that we'll allocate the queue (stealing the task
from the worker that was scheduled) out of order with respect to other
workers that were added to the wiphy workqueue, but that's fine. We are
flushing anyways...

Allocate the queue in-place if needed (from mac_flush()) and remove the
wiphy_work_flush().

Change-Id: Ib302dd5fd1c82244c50ccb7c167234149aec5603
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 3d71c518b78763dbe44f684dc90967886636d78f