CHROMIUM: iwl7000: mld: reorder notification handling
Notify the waiter of the notification (if there is one) only after the
notification was registered to the async handling (i.e. it was added to
async_handlers_list)/ or the sync handler was executed.

That way, the wait handler function doesn't have to validate the size
of the notification: it will be validated before adding it to
async_handlers_list/executing it (in case of a sync handler).
If it is not valid it'll be dropped and the waiter
will have a timeout.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I497d839ba75198339a4b6f9c30ebd6f39a53b6b7
iwl7000-tree: 5cee21aa52ee4d3b92c53dd257c9f4e5637a2a6d