CHROMIUM: iwl7000: mld: call iwl_notification_wait_notify after registering the async handler
Notify the waiter of the notification (if there is one) only after the
notification was registered to the async handling (i.e. it was added to
async_handlers_list)
That way, the wait handler function doesn't have to validate the size
of the notification: it will be validated before adding it to
async_handlers_list. If it is not valid it'll be dropped and the waiter
will have a timeout.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I1553b7d8d9472b65ede90acfaa454a28411b9842
iwl7000-tree: 9e908b8383532bfcaec7b1bbf75b56f9414db342