CHROMIUM: iwl7000: mld: improve (scheduled) scan stop race handling
There's an inherent race with mac80211, the driver and the FW,
in that a scan can be stopped by mac80211 and simultaneously
finish/abort in the firmware, especially for scheduled scan.
The driver has to ensure that it doesn't tell mac80211 about
the finish/abort after a new scan starts, otherwise mac80211
is going to assume that the _new_ instance already finished,
rather than the previous one.

We used to fix this by simply flushing the async workers when
stopping scheduled scan. This is problematic, due to flush
causing mac80211 functions to possibly need to be re-entrant,
which they aren't prepared to be.

However, we can fix this differently. Noting that the firmware
cannot send a notification for a given scan after removal, we
can simply use the cancelltation infrastructure to remove any
possible notifications referring to the scan being stopped. In
this way, there cannot be any scan notifications after stop is
successful, and there's no interference with any new scans.

While at it, do some code cleanups in scan and check UIDs from
firmware before using them for array accesses.

Change-Id: I530434fa543bacfab11862042009d547ca1dbc15
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 11c7e3e3cceeefe8d1cf386eb329561e68d6140b