CHROMIUM: iwl7000: mld: improve remain-on-channel race handling
There's an inherent race involving mac80211, the driver and the
firmware, in that a remain-on-chan can be cancelled by mac80211,
and simultaneously expire in the firmware. The driver then has
to ensure that it doesn't handle the expire notification after
mac80211 starts a new remain-on-channel, otherwise mac80211 is
going to assume that the _new_ instance already expired, rather
than the previous one.

We used to fix this by simply flushing the async workers when
creating a new instance, so that any expire notification that
might be queued will call mac80211 at that point, and then not
interfere with the new one. This is problematic, due to flush
causing mac80211 functions to possibly need to be re-entrant,
which they aren't prepared to be.

However, we can fix this differently. Noting that the firmware
cannot send a notification for a given remain-on-channel after
it was cancelled, we can simply use the existing infrastructure
to cancel any possible notifications referring to the remain-
on-channel instance being cancelled. In this way, there cannot
be any remain-on-channel notifications after cancel succeeds,
and there's no interference with any new instances.

Change-Id: I2d0e6c6dd0b20a02b3c426c29101111615a23e6c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4f44a5ea5db59a6ef01f1b5d966e5c429034e31a