CHROMIUM: iwl7000: mld: always add STA to FW in restart
In iwl_mld_update_link_stas, if a mld_link_sta object that is supposed
to be added - already exist, we skip its addition with a warning.
But if we are in restart, even if the iwl_mld_link_sta exists we should
still add it to the FW, otherwise we will get an assert from the FW in
the next command (TLC of a STA that the FW isn't aware of).

Fix this by always calling iwl_mld_add_link_sta from
iwl_mld_update_link_stas, while changing iwl_mld_add_link_sta to be more
careful:
Even if the iwl_mld_link_sta exists in the driver, add it to the FW.
Warn where it isn't expected to exist.
This will fix the FW assert.

Also, detect if the link STA already exists in the FW, then this is a bug.
Warn and return without adding the STA to the FW, to avoid a FW assert.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ibcf7b8edf3a7be90069d1876c97d74e080feb039
iwl7000-tree: 285943eac41a7c9951128b0e5c9d3e32b3685326