FROMGIT: wifi: iwlwifi: support BIOS override for 5G9 in CA also in LARI version 8
commit b1e8102a4048003097c7054cbc00bbda91a5ced7 upstream.

Commit 6b3e87cc0ca5 ("iwlwifi: Add support for LARI_CONFIG_CHANGE_CMD
cmd v9")
added a few bits to iwl_lari_config_change_cmd::oem_unii4_allow_bitmap
if the FW has LARI version >= 9.
But we also need to send those bits for version 8 if the FW is capable
of this feature (indicated with capability bits)
Add the FW capability bit, and set the additional bits in the cmd when
the version is 8 and the FW capability bit is set.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Link: https://patch.msgid.link/20241226174257.dc5836f84514.I1e38f94465a36731034c94b9811de10cb6ee5921@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit fd82d29c4cfb4767167823ffac6825525c13d2cc
 https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git linux-6.6.y)

BUG=b:391894568
TEST=Verify the patch on all Intel Wi-Fi chips on ChromeOS Flex

Change-Id: I4d9a332a96db8f72007db929352473460e0d859e
Signed-off-by: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6418782
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>