CHROMIUM: iwl7000: mld: allow to disable the sensitivity calibration for WFA
WFA sniffers work in clean environment and we want to ignore packets
received with a very low RSSI. This will allow us to focus on the right
data. Of course, most of the work is in the firmware, but add a
configuration hook to disable the sensitivity calibration which will
tell the firmware to use the low sensitivity mode in case a sniffer MAC
is added.

Change-Id: I0ed29d461504aea00519bdd2e65430405f169238
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: ec79d4ec57f91465d71b4a9e47fba73bcc3c18e6