CHROMIUM: iwl7000: mld: PHC: implement the adjfine() callback
Implement the adjfine() callback which allows to modify the clock
frequency. Since we cannot really modify the GP2 clock frequency,
the adjustments are done in the driver. This is done by saving a
timestamp of the frequency adjustment and the new scale. When the
clock needs to report the current time (e.g. when gettime() is
called), it calculates the time that has passed since the last
scale update, and scales it according to the configured frequency.

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Change-Id: Ifae56316b6cac466fe0f929bcef97fd04fb8878c
iwl7000-tree: 07d6a35f0138a369895d52d3ee4eed9524da5c95