CHROMIUM: iwl7000: mld: check carefully if to activate a link
When switching links, FW requires that the (link) STA will be added
before the link is activated.
Checking whether a link can be activated or not is done by checking if
the mld_link_sta exists for this link.
But in some cases, like in reconfig, there might be an inconsistency
between the driver and the FW. In that case, we currently activate the
link before adding the (link) STA, leading to a FW assert.
Fix this by also checking if the (link) STA is in FW.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I1fd43148eecfff8a361b11eabe1859dfff258b1c
iwl7000-tree: 4d19f7de56c3a8148b3753c27864a4c5c5554a48