FROMGIT: Revert "PCI: Use preserve_config in place of pci_flags"
This reverts commit 3e221877dd92dfeccc840700868e7fef2675181b which is
commit 7246a4520b4bf1494d7d030166a11b5226f6d508 upstream.

This patch causes a regression in cuttlefish/crossvm boot on arm64.

The patch was part of a series that when applied will not cause a regression
but this patch was backported to the 6.6 branch by itself.

The other patches do not apply cleanly to the 6.6 branch.

Signed-off-by: Terry Tritton <terry.tritton@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit bcb9678b1c4f70b9545655b6d41bdb6d0462f281
https://kernel.googlesource.com/pub/scm/linux/kernel/git/stable/linux.git linux-6.6.y)

BUG=b:396689471
TEST=Build tael guest kernel, build arm termina-dlc, boot Crostini

Change-Id: Ic50968425dae2bac75a22d9dc26895a3386bdfbb
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6270357
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Zihan Chen <zihanchen@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Zihan Chen <zihanchen@google.com>