CHROMIUM: iwl7000: mac80211: fix U-APSD check in ML reconfiguration
If U-APSD isn't enabled by us, then IEEE80211_STA_UAPSD_ENABLED
won't be set, but the AP can still support it in that case. Only
require U-APSD from the AP if we enabled it, don't require it to
be disabled on the AP if we didn't.

Change-Id: I01959e448c6a2a3e8bc5d561bbae9e8d2cca616a
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 56aa2f44cea9a12ea79edb2e153ecd428d56a6fc