UPSTREAM: nfsd: restore callback functionality for NFSv4.0
A recent patch inadvertently broke callbacks for NFSv4.0.

In the 4.0 case we do not expect a session to be found but still need to
call setup_callback_client() which will not try to dereference it.

This patch moves the check for failure to find a session into the 4.1+
branch of setup_callback_client()

Fixes: 1e02c641c3a4 ("NFSD: Prevent NULL dereference in nfsd4_process_cb_update()")
Signed-off-by: NeilBrown <neilb@suse.de>
Reviewed-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
(cherry picked from commit 7917f01a286ce01e9c085e24468421f596ee1a0c)

BUG=None
TEST=None

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Ifad4659cf210ad1ccad2e3ace2aad78888a63799
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6186218
Reviewed-by: Divya Chauhan <chauhandivya@google.com>
Reviewed-by: Nitin Lakra <nitinlakra@google.com>
Commit-Queue: Divya Chauhan <chauhandivya@google.com>