CHROMIUM: iwl7000: mld: avoid multiple MLO scans
In case of missed beacon, or low RSSI indication in the statistics,
we trigger a MLO scan, followed by a link selection, to check whether
we can switch to a better link.

But we can have repeated missed beacons notifications (if we decided not
to switch, or if the switch wasn't executed yet). In that case we
currently trigger an MLO scan again and again.

Use iwl_mld_trigger_link_selection that before scanning, checks if the
latest scan was recent enough and in that case reuses the result of the
latest link selection.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Iaf56588a4fb5bb703f7c6a228ee223aed4236d0d
iwl7000-tree: cf88f0a6eec5ab6eb1d4957dd30edaedbf8156e8