CHROMIUM: iwl7000: mld: delete unneeded logic
iwl_mld_unblocked_emlsr is called when the last EMLSR blocker is
removed. It checks if the last exit from EMLSR was recent,
and if so - retruns immediately to EMLSR without link selection.

But this logic was replaced with the one in
iwl_mld_trigger_link_selection, so it is not needed. Remove it.

Change-Id: I9b1ff0bffb427bbe6f5a4361b591041e099ee093
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 0c0a5028e7b17e74d1b7aa2873f2c353337bc2d8