CHROMIUM: iwl7000: mld: don't try to select 2 links if EMLSR is blocked
Currently the link selection logic tries to select 2 links regardless of
whether EMLSR is blocked or not. If it is, it will activate only the
selected primary, but the full selection result will be stored in
mld_vif::emlsr::selected_links.

While this can save us from running the link selection if the last EMLSR
exit was recent, this complicates the code.
(And will get more complicated in the coming patches)

Change the code to skip the 2 links selection and to store the one link
in the result.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ied29fea2974cab85700d358b8096d255646ede05
iwl7000-tree: c645432e7425c6c1812c4cfa9a3fed74b597d7d4