CHROMIUM: iwl7000: w/a FW SMPS mode selection
The FW is now responsible of determining the SMPS mode.
If the user disabled power save in a certain vif, we send the vif-level
power command to clear out the POWER_FLAGS_POWER_MANAGEMENT_ENA_MSK bit
for that vif.
But erroneously, the FW checks DEVICE_POWER_FLAGS_POWER_SAVE_ENA_MSK in
the device-level command to determine the SMPS mode.

To W/A this, send also the device-level command when the power save of a
vif changes, and disable power save if there is any vif that has power
save disabled.

Signed-off-by: Daniel Gabay <daniel.gabay@intel.com>
Change-Id: I2c793ff1fc2a6779a95faaee1ded348100fd97f1
iwl7000-tree: a1e897d4b2dca486b126a1ad5e45edc07aa823c6