CHROMIUM: iwl7000: fix debug actions order
The order of actions taken for debug was implemented incorrectly.
Now we implemented the dump split and do the FW reset only in the
middle of the dump (rather than the FW killing itself on error.)
As a result, some of the actions taken when applying the config
will now crash the device, so we need to fix the order.

Change-Id: I40632c48e2a67a8aca05def572a934b88ce7934b
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 7344a0ab1008ba910ad3a2bd6d55fbd7fe9b563a